packages/poetry-python/pyproject.toml
[tool.poetry]
name = "poetry-python"
version = "0.1.0"
description = ""
authors = ["Daniel Tatarkin <danieltatarkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
boto3 = "^1.28.73"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
